Split a test file into multiple files by function executor
